relevate_web_app/requirements.txt
alabaster==0.7.9
awsebcli==3.9.0
Babel==2.3.4
beautifulsoup4==4.5.1
blessed==1.9.5
boto==2.42.0
botocore==1.4.92
cement==2.8.2
colorama==0.3.7
Django==1.10.1
django-braces==1.9.0
django-storages==1.5.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
imagesize==0.7.1
itsdangerous==0.24
Jinja2==2.9.4
jmespath==0.9.0
MarkupSafe==0.23
mysqlclient==1.3.7
olefile==0.44
pathspec==0.5.0
Pillow==4.0.0
Pygments==2.1.3
python-dateutil==2.6.0
python-http-client==2.2.1
pytz==2016.10
PyYAML==3.12
requests==2.9.1
selenium==3.0.1
semantic-version==2.5.0
sendgrid==3.4.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.1
tabulate==0.7.5
wcwidth==0.1.7
websocket-client==0.40.0
django-jquery-js==3.1.1
pyPdf==1.13
PyPDF2==1.26.0
wand==0.4.4
numpy==1.12.0
djangorestframework==3.6.2